refactoring the processor code to use a return struct instead of std::expected: code is now much cleaner and concise at about the same run-time cost!
Also, I like to have the (b0rked) output string *value* to co-exist with the output *error*, even if the output string value is *empty*. It does not have to be like that, as it adds nothing of value to this particular bit of code, but it's just... an opinion. Heck, I can redo this as a return value *union* if I want to mimic std::expected more closely... Doing it as a *return struct* however does greatly simplify copy/move implementation and both std::string and ErrorInfo (pointer) are, to my mind at least, pretty heavy already, so wrapping in struct vs union: *meh*. Just ride with `struct` and take the extra cost of a single pointer copy when you're in the regular (fast) value-good path. I care much less about the error-path run-time performance so schlepping a surplus std::string around is completely fine, to me.

This takes care of the shite due to LWG/3938 in combination with our use of std::expected<..., std::unique_ptr<...>> which is one example of what LWG/3938 (https://cplusplus.github.io/LWG/issue3938) is about. Our current sanitation-driver.cpp source file still shows all the (MSVC2026) errors we got while we tried to be 'modern' in our implementation.